+++ tests/frameservers/tui_text/tui_test.c
@@ -84,7 +84,7 @@ static void on_mouse(struct tui_context* c,
 }
 
 static void on_key(struct tui_context* c, uint32_t keysym,
-	uint8_t scancode, uint8_t mods, uint16_t subid, void* t)
+	uint8_t scancode, uint16_t mods, uint16_t subid, void* t)
 {
 	trace("unknown_key(%"PRIu32",%"PRIu8",%"PRIu16")", keysym, scancode, subid);
 	struct data_buffer* T = t;
